pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viberank"
version = "0.1.0"
description = "Research codebase for ranking and comparison experiments."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Gaurab Pokharel" }
]
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "jupyter>=1.0",
    "notebook>=7.0",
    "jupyterlab>=4.0",
    "ipykernel>=6.29",
    "tqdm>=4.66",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".venv",
    "data",
    "results",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
]
ignore = [
    "E501",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "^data/",
    "^results/",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]